moving tenancy_licence builder in a separate service class and separating tests for it, saving values a pounds instead of pennies in database and other small changes
